+++ src/main/java/com/sample/urlshortener/service/EncodeService.java
@@ -1,6 +1,6 @@
 package com.sample.urlshortener.service;
 
 public interface EncodeService {
-    String encode(int n);
+    String encode(long n);
     int decode(String s);
 }
